third_party/WebKit/LayoutTests/inspector/sources/debugger/open-close-open-expected.txt

Tests that scripts panel displays resource content correctly after the open - close - open sequence. Bug 56747


Running: testSourceFrameContent
==Source frame contents start==




<script>

function test()
{
    InspectorTest.evaluateInPage("frontendReopeningCount", function(result) {
        if (result._description === "0")
            InspectorTest.evaluateInPage("reopenFrontend()")
        else {
            InspectorTest.runDebuggerTestSuite([
                function testSourceFrameContent(next)
                {
                    InspectorTest.showScriptSource("open-close-open.html", didShowScriptSource);

                    function didShowScriptSource(sourceFrame)
                    {
                        InspectorTest.dumpSourceFrameContents(sourceFrame);
                        next();
                    }
                }
            ]);
        }
    });
}

</script>
==Source frame contents end==
